modified flow for pool shares
