+++ src/components/Profile/History/PoolShares/Liquidity.module.css
@@ -11,25 +11,6 @@
   line-height: 1;
 }
 
-.poolSharesTable [role='gridcell'] {
-  align-items: flex-start;
-  margin: calc(var(--spacer) / 2) 0;
-}
-
-.poolSharesTable [class*='AssetListTitle_title'] {
-  line-height: 0 !important;
-}
-
-.poolSharesTable [class*='Token_token'] div {
-  color: var(--color-secondary);
-}
-
-@media (min-width: 30rem) {
-  .poolSharesTable [class*='AssetListTitle_title'] {
-    line-height: 0 !important;
-  }
-}
-
 .userLiquidity {
   display: flex;
   flex-direction: column;
